Changed the x and y attribute in the Entity class from ints to doubles and did the necessary changes in the rest of the code.

+++ OOP-project/src/main/java/com/group23/app/Model/Entity.java
@@ -1,20 +1,20 @@
 package com.group23.app.Model;
 
 public abstract class Entity {
-    protected int x,y;
+    protected double x,y;
     protected int width,height;
 
-    Entity(int x, int y, int width, int height) {
+    Entity(double x, double y, int width, int height) {
         this.x = x;
         this.y = y;
         this.width = width;
         this.height = height;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
-    public int getY() {
+    public double getY() {
         return y;
     }
     public int getWidth() {
